Moved the data_coordinator class up.
