+++ include/lbann/data_ingestion/CMakeLists.txt
@@ -25,6 +25,7 @@
 ################################################################################
 # Add the headers for this directory
 set_full_path(THIS_DIR_HEADERS
+  data_coordinator.hpp
   data_store_conduit.hpp
   data_reader.hpp
   )
