+++ include/lbann/data_ingestion/coordinator/CMakeLists.txt
@@ -27,7 +27,6 @@
 set_full_path(THIS_DIR_HEADERS
   buffered_data_coordinator.hpp
   buffered_data_coordinator_impl.hpp
-  data_coordinator.hpp
   )
 
 # Propagate the files up the tree
