+++ src/callbacks/dump_minibatch_sample_indices.cpp
@@ -29,7 +29,7 @@
 
 #include "lbann/callbacks/dump_minibatch_sample_indices.hpp"
 #include "lbann/callbacks/callback.hpp"
-#include "lbann/data_ingestion/coordinator/data_coordinator.hpp"
+#include "lbann/data_ingestion/data_coordinator.hpp"
 #include "lbann/execution_algorithms/sgd_execution_context.hpp"
 #include "lbann/models/model.hpp"
 #include "lbann/trainers/trainer.hpp"
