+++ src/data_ingestion/CMakeLists.txt
@@ -25,6 +25,7 @@
 ################################################################################
 # Add the source files for this directory
 set_full_path(THIS_DIR_SOURCES
+  data_coordinator.cpp
   data_store_conduit.cpp
   data_reader.cpp
 )
