+++ src/data_ingestion/coordinator/CMakeLists.txt
@@ -26,7 +26,6 @@
 # Add the source files for this directory
 set_full_path(THIS_DIR_SOURCES
   buffered_data_coordinator.cpp
-  data_coordinator.cpp
   )
 
 # Propagate the files up the tree
